Factor out TokenDecodeIterator.
BUG=
TEST=unittest
REF_BUG=19010851
REF_CL=85502488
