+++ src/dictionary/system/system_dictionary.cc
@@ -64,6 +64,7 @@
 #include "dictionary/dictionary_token.h"
 #include "dictionary/file/dictionary_file.h"
 #include "dictionary/system/codec_interface.h"
+#include "dictionary/system/token_decode_iterator.h"
 #include "dictionary/system/words_info.h"
 #include "storage/louds/bit_vector_based_array.h"
 #include "storage/louds/louds_trie.h"
@@ -154,149 +155,6 @@ void BuildHiraganaExpansionTable(
     }
   }
 }
-
-// Note that this class is just introduced due to performance reason.
-// Conceptually, it should be in somewhere close to the codec implementation
-// (see comments in Next method for details).
-// However, it is necessary to refactor a bit larger area, especially around
-// codec implementations, to make it happen.
-// Considering the merit to introduce this class, we temporarily put it here.
-// TODO(hidehiko): Move this class into a Codec related file.
-class TokenDecodeIterator {
- public:
-  TokenDecodeIterator(
-      const SystemDictionaryCodecInterface *codec,
-      const LoudsTrie &value_trie,
-      const uint32 *frequent_pos,
-      StringPiece key,
-      const uint8 *ptr)
-      : codec_(codec),
-        value_trie_(&value_trie),
-        frequent_pos_(frequent_pos),
-        key_(key),
-        state_(HAS_NEXT),
-        ptr_(ptr),
-        token_info_(nullptr) {
-    key.CopyToString(&token_.key);
-    NextInternal();
-  }
-
-  ~TokenDecodeIterator() {
-  }
-
-  const TokenInfo& Get() const { return token_info_; }
-  bool Done() const { return state_ == DONE; }
-
-  void Next() {
-    DCHECK_NE(state_, DONE);
-    if (state_ == LAST_TOKEN) {
-      state_ = DONE;
-      return;
-    }
-    NextInternal();
-  }
-
- private:
-  enum State {
-    HAS_NEXT,
-    LAST_TOKEN,
-    DONE,
-  };
-
-  void NextInternal() {
-    // Reset token_info with preserving some needed info in previous token.
-    int prev_id_in_value_trie = token_info_.id_in_value_trie;
-    token_info_.Clear();
-    token_info_.token = &token_;
-
-    // Do not clear key in token.
-    token_info_.token->attributes = Token::NONE;
-
-    // This implementation is depending on the internal behavior of DecodeToken
-    // especially which fields are updated or not. Important fields are:
-    // Token::key, Token::value : key and value are never updated.
-    // Token::cost : always updated.
-    // Token::lid, Token::rid : updated iff the pos_type is neither
-    //   FREQUENT_POS nor SAME_AS_PREV_POS.
-    // Token::attributes : updated iff the value is SPELLING_COLLECTION.
-    // TokenInfo::id_in_value_trie : updated iff the value_type is
-    //   DEFAULT_VALUE.
-    // Thus, by not-reseting Token instance intentionally, we can skip most
-    //   SAME_AS_PREV operations.
-    // The exception is Token::attributes. It is not-always set, so we need
-    // reset it everytime.
-    // This kind of structure should be packed in the codec or some
-    // related but new class.
-    int read_bytes;
-    if (!codec_->DecodeToken(ptr_, &token_info_, &read_bytes)) {
-      state_ = LAST_TOKEN;
-    }
-    ptr_ += read_bytes;
-
-    // Fill remaining values.
-    switch (token_info_.value_type) {
-      case TokenInfo::DEFAULT_VALUE: {
-        token_.value.clear();
-        LookupValue(token_info_.id_in_value_trie, &token_.value);
-        break;
-      }
-      case TokenInfo::SAME_AS_PREV_VALUE: {
-        DCHECK_NE(prev_id_in_value_trie, -1);
-        token_info_.id_in_value_trie = prev_id_in_value_trie;
-        // We can keep the current value here.
-        break;
-      }
-      case TokenInfo::AS_IS_HIRAGANA: {
-        token_.value = token_.key;
-        break;
-      }
-      case TokenInfo::AS_IS_KATAKANA: {
-        if (!key_.empty() && key_katakana_.empty()) {
-          Util::HiraganaToKatakana(key_, &key_katakana_);
-        }
-        token_.value = key_katakana_;
-        break;
-      }
-      default: {
-        LOG(DFATAL) << "unknown value_type: " << token_info_.value_type;
-        break;
-      }
-    }
-
-    if (token_info_.accent_encoding_type == TokenInfo::EMBEDDED_IN_TOKEN) {
-      token_.value.append(1, '_')
-                  .append(Util::StringPrintf("%d", token_info_.accent_type));
-    }
-
-    if (token_info_.pos_type == TokenInfo::FREQUENT_POS) {
-      const uint32 pos = frequent_pos_[token_info_.id_in_frequent_pos_map];
-      token_.lid = pos >> 16;
-      token_.rid = pos & 0xffff;
-    }
-  }
-
-  void LookupValue(int id, string *value) const {
-    char buffer[LoudsTrie::kMaxDepth + 1];
-    const StringPiece encoded_value = value_trie_->RestoreKeyString(id, buffer);
-    codec_->DecodeValue(encoded_value, value);
-  }
-
-  const SystemDictionaryCodecInterface *codec_;
-  const LoudsTrie *value_trie_;
-  const uint32 *frequent_pos_;
-
-  const StringPiece key_;
-  // Katakana key will be lazily initialized.
-  string key_katakana_;
-
-  State state_;
-  const uint8 *ptr_;
-
-  TokenInfo token_info_;
-  Token token_;
-
-  DISALLOW_COPY_AND_ASSIGN(TokenDecodeIterator);
-};
 
 inline const uint8 *GetTokenArrayPtr(const BitVectorBasedArray &token_array,
                                      int key_id) {
